Fix: make a new query when user selects another area
